+++ Blog/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from .models import Blog
+
+# Register your models here.
+admin.site.register(Blog)
