+++ Blog/serializers.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from .models import Blog
+
+class BlogSerializer(serializers.ModelSerializer):
+    class Meta:
+        model= Blog
+        fields = '__all__'
