+++ mhm/urls.py
@@ -13,6 +13,7 @@ route.register("",BlogView, basename='blogview')
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("apis/", include("apis.urls")),
+    path("blog/", include("Blog.urls")),
     path("img/",include(route.urls)),
     path("", TemplateView.as_view(template_name="index.html")),
 ]+ static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
